+++ client/src/app/refactoring/subtoolbar-refactoring/subtoolbar-refactoring.component.html
@@ -1,2 +1,1 @@
-<p-button icon="pi pi-fw pi-cog" styleClass="p-button-text p-button-lg" pTooltip="Refactor" showDelay="1000" tooltipPosition="bottom" (click)="refactoringMenu.toggle($event)"></p-button>
-<p-menu #refactoringMenu [model]="menuitems" [popup]="true"></p-menu>
+<p-button icon="pi pi-fw pi-verified" styleClass="p-button-text p-button-lg" pTooltip="Find smells" showDelay="1000" tooltipPosition="bottom" (click)="analyse()"></p-button>
